Multilingual_Code_Explanation_API/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
pypdf>=3.0.0
chardet>=4.0.0
python-dotenv>=0.19.0
ibm-watsonx-ai>=0.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0
reportlab>=4.0.0
pytest>=6.2.5
pytest-asyncio>=0.18.0
httpx>=0.23.0
aiohttp>=3.8.0 